+++ design-system/code/__perf__/default.tsx
@@ -2,4 +2,4 @@ import React from 'react';
 
 import { Code } from '../src';
 
-export default () => <Code>bolt changeset</Code>;
+export default () => <Code>yarn changeset</Code>;