+++ design-system/tokens/scripts/style-dictionary/formatters/format-rename-mapper.tsx
@@ -23,7 +23,7 @@ const formatter: Format['formatter'] = ({ dictionary }) => {
  * 1. Mark the old token's 'state' as deprecated
  * 2. Add a 'rename' attribute to the token with the value 'my.new.token'
  * 3. Create a new token matching the token above: 'my.new.token'
- * 4. Run 'bolt build tokens' to have you changes reflected in this map
+ * 4. Run 'yarn build tokens' to have you changes reflected in this map
  * 5. eslint and other tools will now use this to automate replacing tokens
  *
  * These changes will then be picked up by our tooling, which will attempt to