+++ editor/editor-core/src/plugins/code-block/__tests__/__fixtures__/table-code.ts
@@ -24,7 +24,7 @@ export const tableCode = {
                   content: [
                     {
                       type: 'text',
-                      text: 'bolt changeset',
+                      text: 'yarn changeset',
                       marks: [
                         {
                           type: 'code',