+++ smart-experiences/smart-hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/smart-hooks
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.0.2
 
 ### Patch Changes